Feature/order

+++ src/main/java/com/tripleseven/orderapi/business/point/PointService.java
@@ -3,9 +3,9 @@ package com.tripleseven.orderapi.business.point;
 import com.tripleseven.orderapi.dto.pointhistory.PointHistoryResponseDTO;
 
 public interface PointService {
-    PointHistoryResponseDTO createPointHistoryForPaymentSpend(Long memberId, int usedPoint, Long orderGroupId);
+    PointHistoryResponseDTO createPointHistoryForPaymentSpend(Long memberId, long usedPoint, Long orderGroupId);
 
-    PointHistoryResponseDTO createPointHistoryForPaymentEarn(Long memberId, int usedMoney, Long orderGroupId);
+    PointHistoryResponseDTO createPointHistoryForPaymentEarn(Long memberId, long usedMoney, Long orderGroupId);
 
     PointHistoryResponseDTO createRegisterPointHistory(Long memberId);
 }
